+++ keras/tests/integration_test.py
@@ -26,7 +26,7 @@ from keras.testing_infra import test_combinations
 from keras.testing_infra import test_utils
 from keras.layers.rnn import legacy_cells
 from keras.legacy_tf_layers import base as base_layer
-from keras.utils import all_utils as utils
+from keras import utils
 
 
 class KerasIntegrationTest(test_combinations.TestCase):
