--- keras/utils/all_utils.py
@@ -1,41 +1,0 @@
-# Copyright 2019 The TensorFlow Authors. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-# ==============================================================================
-"""Public Keras utilities.
-
-This module is used as a shortcut to access all the symbols. Those symbols was
-exposed under __init__, and was causing some hourglass import issue.
-"""
-
-# pylint: disable=unused-import
-from keras.utils.data_utils import GeneratorEnqueuer
-from keras.utils.data_utils import get_file
-from keras.utils.data_utils import OrderedEnqueuer
-from keras.utils.data_utils import Sequence
-from keras.utils.data_utils import SequenceEnqueuer
-from keras.utils.generic_utils import custom_object_scope
-from keras.utils.generic_utils import CustomObjectScope
-from keras.utils.generic_utils import deserialize_keras_object
-from keras.utils.generic_utils import get_custom_objects
-from keras.utils.generic_utils import Progbar
-from keras.utils.generic_utils import serialize_keras_object
-from keras.utils.image_dataset import image_dataset_from_directory
-from keras.utils.layer_utils import get_source_inputs
-from keras.utils.np_utils import normalize
-from keras.utils.np_utils import to_categorical
-from keras.utils.text_dataset import text_dataset_from_directory
-from keras.utils.tf_utils import set_random_seed
-from keras.utils.timeseries_dataset import timeseries_dataset_from_array
-from keras.utils.vis_utils import model_to_dot
-from keras.utils.vis_utils import plot_model
